+++ src/borg/archiver/completion_cmd.py
@@ -715,6 +715,7 @@ class CompletionMixIn:
             "buzhash64,19,23,21,4095,2",
             "fastcdc,19,23,21,2",
             "rabin-aes,19,23,21,2",
+            "goldilocks-aes,19,23,21,2",
         ]
         chunker_params_choices_str = " ".join(chunker_params_choices)